added characters and character Paths filler
