+++ app/controllers/westerosController/character.js
