+++ app/controllers/westerosController/characterLocations.js
