+++ app/controllers/westerosController/characterPaths.js
