+++ app/controllers/westerosController/event.js
